refactor: [RABBIT-60] trade 컴포넌트 분리 및 리팩토링
