+++ .gitignore
@@ -105,3 +105,5 @@ dist
 
 # Package files
 package-lock.json
+
+.DS_Store
