+++ src/app/_shared/service/OAuthLogin.ts
@@ -1,10 +1,10 @@
-
+const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL;
 
 function OAuthLogin(
   provider: "google" | "github" | "kakao" | "naver"
 ) {
   try {
-    const url = `http://localhost:8080/login/${provider}`;
+    const url = `${API_BASE_URL}/login/${provider}`;
     window.location.assign(url);
   } catch (error) {
     console.error("로그인 실패:", error);
@@ -15,7 +15,7 @@ function OAuthLogin(
 
 function Logout() {
   try {
-    const url = `http://localhost:8080/auth/logout`;
+    const url = `${API_BASE_URL}/auth/logout`;
     window.location.assign(url);
   } catch (error) {
     console.error("로그아웃 실패:", error);
