+++ src/app/globals.css
@@ -30,6 +30,7 @@ html,
 body {
     max-width: 100vw;
     overflow-x: hidden;
+    color: #000;
 }
 
 body {
